split rendering into separate module
