+++ lib/notifications/github.bot.js
@@ -4,12 +4,10 @@ const Promise = require('bluebird')
 const clone = require('101/clone')
 const isEmpty = require('101/is-empty')
 const isString = require('101/is-string')
-const keypather = require('keypather')()
 const pluck = require('101/pluck')
-const runnableHostname = require('runnable-hostname')
 const GitHub = require('models/github')
+const GitHubBotMessage = require('notifications/github.bot.message')
 const tracker = require('models/tracker')
-const utils = require('models/utils')
 const logger = require('logger').child({ module: 'notifications/github-bot' })
 
 function getToken () {
@@ -202,7 +200,7 @@ class GitHubBot {
           log.error({ err: err }, 'error looking for a comment')
           return cb(err)
         }
-        const newMessage = this._render(gitInfo, instance, isolatedInstances)
+        const newMessage = GitHubBotMessage.render(gitInfo, instance, isolatedInstances)
         log.info({ message: newMessage }, 'rendered message')
         if (!newMessage) {
           log.info('message is null, do not send anything')
@@ -264,112 +262,6 @@ class GitHubBot {
           })
         })
       })
-  }
-
-  _renderFooter (isolatedInstances) {
-    let isolatedLinks = this._renderIsolatedInstance(isolatedInstances)
-    if (isolatedLinks.length === 0) {
-      return '\n<sub>*From [Runnable](http://runnable.com)*</sub>'
-    }
-    return '\n' + isolatedLinks + '*— From [Runnable](http://runnable.com)*</sub>'
-  }
-
-  _renderIsolatedInstance (isolatedInstances) {
-    let message = ''
-    if (isolatedInstances && !isEmpty(isolatedInstances)) {
-      message += '<sub>Related containers: '
-      const linksList = isolatedInstances.map((instance) => {
-        const instanceUrl = process.env.WEB_URL + '/' + instance.owner.username + '/' +
-            instance.name
-        const readableInstanceName = this._cleanupInstanceName(instance.name)
-        const link = this._createMarkdownLink(readableInstanceName, instanceUrl)
-        const statusIcon = this._renderStatusIcon(utils.instanceState(instance))
-        return statusIcon + ' ' + link
-      })
-      message += linksList.join(' ')
-    }
-    return message
-  }
-
-  _renderStatusIcon (state) {
-    if (state === 'running') {
-      return '<img src="https://s3-us-west-1.amazonaws.com/runnable-design/status-green.svg" title="Running" width="9" height="9">'
-    }
-    if (state === 'stopped') {
-      return '<img src="https://s3-us-west-1.amazonaws.com/runnable-design/status-gray.svg" title="Stopped" width="9" height="9">'
-    }
-    if (state === 'building') {
-      return '<img src="https://s3-us-west-1.amazonaws.com/runnable-design/status-orange.svg" title="Building" width="9" height="9">'
-    }
-    return '<img src="https://s3-us-west-1.amazonaws.com/runnable-design/status-red.svg" title="Failed" width="9" height="9">'
-  }
-
-  _render (gitInfo, instance, isolatedInstances) {
-    const opts = {
-      shortHash: instance.shortHash,
-      instanceName: instance.name,
-      branch: gitInfo.branch,
-      ownerUsername: instance.owner.username,
-      masterPod: instance.masterPod,
-      userContentDomain: process.env.USER_CONTENT_DOMAIN
-    }
-    // each message has metadata specified as html comments
-    let message = '<!--'
-    message += 'instanceId:' + instance._id
-    message += '-->'
-
-    message += 'Deployed '
-    const statusIcon = this._renderStatusIcon(gitInfo.state)
-    message += statusIcon + ' '
-    const containerUrl = process.env.CONTAINER_URL_PROTOCOL + '://' + runnableHostname.direct(opts) + this._defaultPort(instance)
-    message += this._createMarkdownLink(instance.name, containerUrl)
-    message += ' to '
-    const instanceUrl = process.env.WEB_URL + '/' + instance.owner.username + '/' + instance.name
-    message += this._createMarkdownLink('your environment', instanceUrl)
-    message += this._renderFooter(isolatedInstances)
-    return message
-  }
-
-  _defaultPort (instance) {
-    const ports = this._instancePorts(instance)
-    if (ports.length > 0) {
-      if (ports.indexOf('80') === -1) {
-        return ':' + ports[0]
-      }
-    }
-    return ''
-  }
-
-  _instancePorts (instance) {
-    const portsObj = keypather.get(instance, 'containers[0].ports')
-    if (!portsObj) {
-      return []
-    }
-    return Object.keys(portsObj).map((port) => {
-      return port.split('/')[0]
-    })
-  }
-
-  _cleanupInstanceName (instanceName) {
-    // if fullname prefixed by shortHash and -- we need to cleaned it up
-    // drop only first `--`
-    const fullName = instanceName
-    const splitByPrefix = fullName.split('--')
-    let shortName = splitByPrefix[0]
-    if (splitByPrefix.length > 1) {
-      splitByPrefix.shift()
-      shortName = splitByPrefix.join('--')
-    }
-    return shortName
-  }
-  /**
-   * Produce a markdown-formatted link
-   * @param {String} link title
-   * @param {String} link url
-   * @return {String}
-   */
-  _createMarkdownLink (title, url) {
-    return '[' + title + '](' + url + ')'
   }
 }
